Create ability to build rpm in docker

+++ rpmbuild.sh
@@ -5,13 +5,22 @@ SPECFILE="$(dirname "$(realpath "$0")")/finddata.spec"
 VERSION=$(grep ^version pyproject.toml  | cut -d " " -f 3 | sed 's/\"//g')
 echo "version in pyproject.toml is ${VERSION}"
 
-# create the tarball
-echo "building sdist..."
-pixi run build-sdist || exit 127
-
+# create the source tarball
+if [ "$(command -v pixi)" ]; then
+    # this runs outside of docker where pixi exists
+    echo "building sdist..."
+    pixi run build-sdist || exit 127
+else
+    # must be inside of docker
+    echo "assuming sdist already exists"
+fi
 
 # setup rpm directories for building - renames the tarball
 TARBALL_SRC="finddata-${VERSION}.tar.gz" # created
+if [ ! -f dist/"${TARBALL_SRC}" ]; then
+    echo "Failed to find source tarball: dist/${TARBALL_SRC}"
+    exit 127
+fi
 mkdir -p "${HOME}"/rpmbuild/SOURCES
 cp dist/"${TARBALL_SRC}" "${HOME}"/rpmbuild/SOURCES/"${TARBALL_SRC}" || exit 127
